chore(api): edit request so that Stored Procedure 'x' can be used for filtering orders
Modify the parameters so that they represent the correct values given by the expected SP. This does not have the correct SP yet and is added in a later version.

Add GetUsers which retrieves users from the database. This does not have the correct SP yet and is added in a later version.